Revert "EVG-20360: Check user permissions in RemoveItemFromCommitQueue resolver (#6946)"
This reverts commit 16070bea0babee794fc4389a589771aff39f6f95.
