+++ graphql/mutation_resolver.go
@@ -1265,9 +1265,8 @@ func (r *mutationResolver) UpdateUserSettings(ctx context.Context, userSettings 
 
 // RemoveItemFromCommitQueue is the resolver for the removeItemFromCommitQueue field.
 func (r *mutationResolver) RemoveItemFromCommitQueue(ctx context.Context, commitQueueID string, issue string) (*string, error) {
-	usr := mustHaveUser(ctx)
-
-	result, err := data.FindAndRemoveCommitQueueItem(ctx, commitQueueID, issue, usr.DisplayName(), fmt.Sprintf("removed by user '%s'", usr.DisplayName()))
+	username := gimlet.GetUser(ctx).DisplayName()
+	result, err := data.FindAndRemoveCommitQueueItem(ctx, commitQueueID, issue, username, fmt.Sprintf("removed by user '%s'", username))
 	if err != nil {
 		return nil, InternalServerError.Send(ctx, fmt.Sprintf("error removing item %s from commit queue %s: %s",
 			issue, commitQueueID, err.Error()))
