+++ graphql/resolver.go
@@ -26,10 +26,9 @@ type Resolver struct {
 }
 
 func New(apiURL string) Config {
-	dbConnector := &data.DBConnector{URL: apiURL}
 	c := Config{
 		Resolvers: &Resolver{
-			sc: dbConnector,
+			sc: &data.DBConnector{URL: apiURL},
 		},
 	}
 	c.Directives.RequireDistroAccess = func(ctx context.Context, obj interface{}, next graphql.Resolver, access DistroSettingsAccess) (interface{}, error) {
@@ -199,38 +198,6 @@ func New(apiURL string) Config {
 			return next(ctx)
 		}
 		return nil, Forbidden.Send(ctx, fmt.Sprintf("user does not have permission to access the field '%s' for project with ID '%s'", graphql.GetFieldContext(ctx).Path(), projectId))
-	}
-	c.Directives.RequireCommitQueueItemOwner = func(ctx context.Context, obj interface{}, next graphql.Resolver) (interface{}, error) {
-		usr := mustHaveUser(ctx)
-
-		args, isStringMap := obj.(map[string]interface{})
-		if !isStringMap {
-			return nil, InternalServerError.Send(ctx, "converting mutation args into map")
-		}
-
-		commitQueueId, hasCommitQueueId := args["commitQueueId"].(string)
-		if !hasCommitQueueId {
-			return nil, InputValidationError.Send(ctx, "commit queue id was not provided")
-		}
-
-		issue, hasIssue := args["issue"].(string)
-		if !hasIssue {
-			return nil, InputValidationError.Send(ctx, "issue was not provided")
-		}
-
-		project, err := data.FindProjectById(commitQueueId, true, false)
-		if err != nil {
-			return nil, InternalServerError.Send(ctx, err.Error())
-		}
-
-		if err = data.CheckCanRemoveCommitQueueItem(ctx, dbConnector, usr, project, issue); err != nil {
-			gimletErr, ok := err.(gimlet.ErrorResponse)
-			if ok {
-				return nil, mapHTTPStatusToGqlError(ctx, gimletErr.StatusCode, err)
-			}
-			return nil, InternalServerError.Send(ctx, err.Error())
-		}
-		return next(ctx)
 	}
 	return c
 }
