+++ graphql/schema/directives.graphql
@@ -17,8 +17,3 @@ directive @requireProjectFieldAccess on FIELD_DEFINITION
 requireDistroAccess is used to restrict view, edit, admin, and create access for distros.
 """
 directive @requireDistroAccess(access: DistroSettingsAccess!) on ARGUMENT_DEFINITION | INPUT_FIELD_DEFINITION
-
-"""
-requireCommitQueueItemOwner is used to restrict changes to commit queue patches.
-"""
-directive @requireCommitQueueItemOwner on ARGUMENT_DEFINITION
