+++ graphql/tests/mutation/removeItemFromCommitQueue/data.json
@@ -1,120 +1,53 @@
 {
   "commit_queue": [
     {
-      "_id": "project_id",
+      "_id": "mongodb-mongo-master",
+      "processing": true,
+      "queue": [
+        {
+          "issue": "5e94c2dfe3c3312519b59480",
+          "patch_id": "5e94c2dfe3c3312519b59480",
+          "source": "diff",
+          "enqueue_time": { "$date": "2020-04-13T19:52:00.182Z" },
+          "modules": null
+        }
+      ],
+      "processing_updated_time": { "$date": "2020-04-13T19:52:10.216Z" }
+    },
+    {
+      "_id": "mongodb-evergreen-spruce",
+      "processing": true,
+      "queue": [
+        {
+          "issue": "1234",
+          "enqueue_time": { "$date": "2020-05-12T19:52:00.182Z" },
+          "modules": null,
+          "source": "PR",
+          "version": "5e6bb9e23066155a993e0f1a"
+        }
+      ],
+      "processing_updated_time": { "$date": "2020-04-13T19:52:10.216Z" }
+    },
+    {
+      "_id": "cli-queue",
       "processing": true,
       "queue": [
         {
           "issue": "5e6bb9e23066155a993effff",
           "patch_id": "5e6bb9e23066155a993effff",
           "enqueue_time": { "$date": "2020-05-12T19:52:00.182Z" }
-        },
-        {
-          "issue": "5ecedafb562343215a7ff297",
-          "patch_id": "5ecedafb562343215a7ff297",
-          "enqueue_time": { "$date": "2020-05-12T19:52:00.182Z" }
         }
       ]
     },
     {
-      "_id": "sandbox_project_id",
+      "_id": "pr-queue",
       "processing": true,
       "queue": [
         {
-          "issue": "5e4ff3abe3c3317e352062e4",
+          "issue": "5678",
           "enqueue_time": { "$date": "2020-05-12T19:52:00.182Z" }
         }
       ]
-    },
-    {
-      "_id": "cq_disabled_project_id",
-      "processing": true,
-      "queue": []
-    }
-  ],
-  "patches": [
-    {
-      "_id": {
-        "$oid": "5e6bb9e23066155a993effff"
-      },
-      "version": "123456789012345678901234",
-      "branch": "main",
-      "author": "testuser"
-    },
-    {
-      "_id": {
-        "$oid": "5ecedafb562343215a7ff297"
-      },
-      "version": "5ecedafb562343215a7ff297",
-      "branch": "main",
-      "author": "not-testuser"
-    },
-    {
-      "_id": {
-        "$oid": "5e4ff3abe3c3317e352062e4"
-      },
-      "version": "5e4ff3abe3c3317e352062e4",
-      "branch": "main",
-      "author": "not-testuser"
-    }
-  ],
-  "project_ref": [
-    {
-      "_id": "project_id",
-      "owner_name": "evergreen-ci",
-      "repo_name": "commit-queue-sandbox",
-      "branch_name": "main",
-      "repo_kind": "github",
-      "enabled": false,
-      "remote_path": "evergreen.yaml",
-      "identifier": "project_id",
-      "display_name": "Project",
-      "hidden": false,
-      "admins": [
-        "admin"
-      ],
-      "commit_queue": {
-        "enabled": true,
-        "merge_method": "squash"
-      }
-    },
-    {
-      "_id": "sandbox_project_id",
-      "owner_name" : "evergreen-ci",
-      "repo_name" : "commit-queue-sandbox",
-      "branch_name": "main",
-      "repo_kind": "github",
-      "enabled": false,
-      "remote_path": "evergreen.yaml",
-      "identifier": "sandbox",
-      "display_name": "Sandbox",
-      "hidden": false,
-      "admins": [
-        "me"
-      ],
-      "commit_queue": {
-        "enabled": true,
-        "merge_method": "squash"
-      }
-    },
-    {
-      "_id": "cq_disabled_project_id",
-      "owner_name": "evergreen-ci",
-      "repo_name": "commit-queue-sandbox",
-      "branch_name": "main",
-      "repo_kind": "github",
-      "enabled": false,
-      "remote_path": "evergreen.yaml",
-      "identifier": "cq_disabled_project_id",
-      "display_name": "Disabled Project",
-      "hidden": false,
-      "admins": [
-        "admin"
-      ],
-      "commit_queue": {
-        "enabled": false,
-        "merge_method": "squash"
-      }
     }
   ]
 }
