+++ graphql/tests/mutation/removeItemFromCommitQueue/queries/cli_remove_item_from_commit_queue.graphql
@@ -1,6 +1,6 @@
 mutation {
   removeItemFromCommitQueue(
-    commitQueueId: "project_id"
+    commitQueueId: "cli-queue"
     issue: "5e6bb9e23066155a993effff"
   )
 }
