--- graphql/tests/mutation/removeItemFromCommitQueue/queries/fail_commit_queue_disabled.graphql
@@ -1,6 +1,0 @@
-mutation {
-  removeItemFromCommitQueue(
-    commitQueueId: "cq_disabled_project_id"
-    issue: "random"
-  )
-}
