+++ graphql/tests/mutation/removeItemFromCommitQueue/queries/pr_remove_item_from_commit_queue.graphql
@@ -1,0 +1,6 @@
+mutation {
+    removeItemFromCommitQueue(
+        commitQueueId: "pr-queue"
+        issue: "5678"
+    )
+}
