+++ graphql/tests/mutation/removeItemFromCommitQueue/results.json
@@ -1,7 +1,7 @@
 {
   "tests": [
     {
-      "query_file": "success_basic_user.graphql",
+      "query_file": "cli_remove_item_from_commit_queue.graphql",
       "result": {
         "data": {
           "removeItemFromCommitQueue": "5e6bb9e23066155a993effff"
@@ -9,51 +9,11 @@
       }
     },
     {
-      "query_file": "success_project_admin.graphql",
+      "query_file": "pr_remove_item_from_commit_queue.graphql",
       "result": {
         "data": {
-          "removeItemFromCommitQueue": "5e4ff3abe3c3317e352062e4"
+          "removeItemFromCommitQueue": "5678"
         }
-      }
-    },
-    {
-      "query_file": "fail_not_authorized.graphql",
-      "result": {
-        "data": {
-          "removeItemFromCommitQueue": null
-        },
-        "errors": [
-          {
-            "message": "401 (Unauthorized): not authorized to perform action on behalf of author",
-            "path": [
-              "removeItemFromCommitQueue",
-              "issue"
-            ],
-            "extensions": {
-              "code": "FORBIDDEN"
-            }
-          }
-        ]
-      }
-    },
-    {
-      "query_file": "fail_commit_queue_disabled.graphql",
-      "result": {
-        "data": {
-          "removeItemFromCommitQueue": null
-        },
-        "errors": [
-          {
-            "message": "400 (Bad Request): commit queue is not enabled for project 'cq_disabled_project_id'",
-            "path": [
-              "removeItemFromCommitQueue",
-              "issue"
-            ],
-            "extensions": {
-              "code": "INPUT_VALIDATION_ERROR"
-            }
-          }
-        ]
       }
     }
   ]
