+++ rest/data/commit_queue.go
@@ -4,19 +4,16 @@ import (
 	"context"
 	"fmt"
 	"net/http"
-	"runtime/debug"
 	"strconv"
 	"strings"
 	"time"
 
 	"github.com/evergreen-ci/evergreen"
 	"github.com/evergreen-ci/evergreen/db"
-	"github.com/evergreen-ci/evergreen/db/mgo/bson"
 	"github.com/evergreen-ci/evergreen/model"
 	"github.com/evergreen-ci/evergreen/model/commitqueue"
 	"github.com/evergreen-ci/evergreen/model/event"
 	"github.com/evergreen-ci/evergreen/model/patch"
-	"github.com/evergreen-ci/evergreen/model/user"
 	restModel "github.com/evergreen-ci/evergreen/rest/model"
 	"github.com/evergreen-ci/evergreen/thirdparty"
 	"github.com/evergreen-ci/evergreen/units"
@@ -579,88 +576,4 @@ func GetAdditionalPatches(patchId string) ([]string, error) {
 		StatusCode: http.StatusNotFound,
 		Message:    errors.Errorf("patch '%s' not found in commit queue", patchId).Error(),
 	}
-}
-
-// CheckCanRemoveCommitQueueItem checks if a patch can be removed from the commit queue by the given user.
-func CheckCanRemoveCommitQueueItem(ctx context.Context, sc Connector, usr *user.DBUser, project *model.ProjectRef, itemId string) error {
-	if !project.CommitQueue.IsEnabled() {
-		return gimlet.ErrorResponse{
-			StatusCode: http.StatusBadRequest,
-			Message:    errors.Errorf("commit queue is not enabled for project '%s'", project.Id).Error(),
-		}
-	}
-
-	if canAlwaysSubmitPatchesForProject(usr, project.Id) {
-		return nil
-	}
-
-	if bson.IsObjectIdHex(itemId) {
-		patch, err := FindPatchById(itemId)
-		if err != nil {
-			return gimlet.ErrorResponse{
-				StatusCode: http.StatusInternalServerError,
-				Message:    errors.Wrapf(err, "finding patch '%s'", itemId).Error(),
-			}
-		}
-		if usr.Id != utility.FromStringPtr(patch.Author) {
-			return gimlet.ErrorResponse{
-				StatusCode: http.StatusUnauthorized,
-				Message:    "not authorized to perform action on behalf of author",
-			}
-		}
-	} else if itemInt, err := strconv.Atoi(itemId); err == nil {
-		pr, err := sc.GetGitHubPR(ctx, project.Owner, project.Repo, itemInt)
-		if err != nil {
-			return gimlet.ErrorResponse{
-				StatusCode: http.StatusInternalServerError,
-				Message:    errors.Wrapf(err, "unable to get pull request info, PR number (%d) may be invalid", itemInt).Error(),
-			}
-		}
-
-		var githubUID int
-		if pr != nil && pr.User != nil && pr.User.ID != nil {
-			githubUID = int(*pr.User.ID)
-		}
-		if githubUID == 0 || usr.Settings.GithubUser.UID != githubUID {
-			return gimlet.ErrorResponse{
-				StatusCode: http.StatusUnauthorized,
-				Message:    "not authorized to perform action on behalf of GitHub user",
-			}
-		}
-	} else {
-		return gimlet.ErrorResponse{
-			StatusCode: http.StatusBadRequest,
-			Message:    "commit queue item is not a valid identifier",
-		}
-	}
-
-	return nil
-}
-
-// canAlwaysSubmitPatchesForProject returns true if the user is a superuser or project admin,
-// or is authorized specifically to patch on behalf of other users.
-func canAlwaysSubmitPatchesForProject(user *user.DBUser, projectId string) bool {
-	if projectId == "" {
-		grip.Error(message.Fields{
-			"message": "projectID is empty",
-			"op":      "middleware",
-			"stack":   string(debug.Stack()),
-		})
-		return false
-	}
-	isAdmin := user.HasPermission(gimlet.PermissionOpts{
-		Resource:      projectId,
-		ResourceType:  evergreen.ProjectResourceType,
-		Permission:    evergreen.PermissionProjectSettings,
-		RequiredLevel: evergreen.ProjectSettingsEdit.Value,
-	})
-	if isAdmin {
-		return true
-	}
-	return user.HasPermission(gimlet.PermissionOpts{
-		Resource:      projectId,
-		ResourceType:  evergreen.ProjectResourceType,
-		Permission:    evergreen.PermissionPatches,
-		RequiredLevel: evergreen.PatchSubmitAdmin.Value,
-	})
 }
