+++ rest/data/commit_queue_test.go
@@ -389,80 +389,3 @@ func TestConcludeMerge(t *testing.T) {
 	require.NoError(t, err)
 	assert.Len(t, queue.Queue, 0)
 }
-
-func TestCheckCanRemoveCommitQueueItem(t *testing.T) {
-	ctx, cancel := context.WithCancel(context.Background())
-	defer cancel()
-
-	require.NoError(t, db.ClearCollections(
-		user.Collection,
-		model.ProjectRefCollection,
-		model.ProjectVarsCollection,
-		evergreen.ScopeCollection,
-		evergreen.RoleCollection,
-		patch.Collection,
-		commitqueue.Collection,
-	))
-
-	basicUser := &user.DBUser{Id: "me"}
-	require.NoError(t, basicUser.Insert())
-
-	projectAdmin := &user.DBUser{Id: "admin"}
-	require.NoError(t, projectAdmin.Insert())
-
-	project := &model.ProjectRef{
-		Id:      "evergreen",
-		Owner:   "evergreen-ci",
-		Repo:    "evergreen",
-		Branch:  "main",
-		Enabled: true,
-		CommitQueue: model.CommitQueueParams{
-			Enabled: utility.TruePtr(),
-		},
-	}
-	require.NoError(t, project.Add(projectAdmin))
-
-	myPatch := patch.Patch{
-		Id:      bson.NewObjectId(),
-		Project: "evergreen",
-		Author:  basicUser.Id,
-	}
-	require.NoError(t, myPatch.Insert())
-
-	otherPatch := patch.Patch{
-		Id:      bson.NewObjectId(),
-		Project: "evergreen",
-		Author:  "other user",
-	}
-	require.NoError(t, otherPatch.Insert())
-
-	mockConnector := &MockGitHubConnector{
-		MockGitHubConnectorImpl: MockGitHubConnectorImpl{},
-	}
-
-	// Basic user can remove their own CLI patch.
-	err := CheckCanRemoveCommitQueueItem(ctx, mockConnector, basicUser, project, myPatch.Id.Hex())
-	require.NoError(t, err)
-
-	// Basic user cannot remove other user's CLI patch.
-	err = CheckCanRemoveCommitQueueItem(ctx, mockConnector, basicUser, project, otherPatch.Id.Hex())
-	require.EqualError(t, err, "401 (Unauthorized): not authorized to perform action on behalf of author")
-
-	// Basic user can remove their own GitHub patch.
-	basicUser.Settings.GithubUser.UID = 1234
-	err = CheckCanRemoveCommitQueueItem(ctx, mockConnector, basicUser, project, "570")
-	require.NoError(t, err)
-
-	// Basic user cannot remove other user's GitHub patch.
-	basicUser.Settings.GithubUser.UID = 4321
-	err = CheckCanRemoveCommitQueueItem(ctx, mockConnector, basicUser, project, "570")
-	require.EqualError(t, err, "401 (Unauthorized): not authorized to perform action on behalf of GitHub user")
-
-	// Project admin can remove other user's CLI patch.
-	err = CheckCanRemoveCommitQueueItem(ctx, mockConnector, projectAdmin, project, otherPatch.Id.Hex())
-	require.NoError(t, err)
-
-	// Project admin can remove other user's GitHub patch.
-	err = CheckCanRemoveCommitQueueItem(ctx, mockConnector, projectAdmin, project, "570")
-	require.NoError(t, err)
-}
