+++ testdata/local/commit_queue.json
@@ -1,3 +1,4 @@
 {"_id":"mongodb-mongo-master","processing":true,"queue":[{"issue":"5e94c2dfe3c3312519b59480","source":"diff","version":"5e94c2dfe3c3312519b59480","enqueue_time":{"$date":"2020-04-13T19:52:00.182Z"},"modules":null}],"processing_updated_time":{"$date":"2020-04-13T19:52:10.216Z"}}
 {"_id":"mongodb-mongo-test","processing":true,"queue":[{"issue":"5e4ff3abe3c3317e352062e4","source":"diff","version":"5e4ff3abe3c3317e352062e4","enqueue_time":{"$date":"2020-04-13T19:52:00.182Z"},"modules":null}],"processing_updated_time":{"$date":"2020-04-13T19:52:10.216Z"}}
 {"_id":"evergreen","processing":true,"queue":[{"issue":"5e4ff3abe3c3317e352062e4","source":"diff","version":"5e4ff3abe3c3317e352062e4","enqueue_time":{"$date":"2020-04-13T19:52:00.182Z"},"modules":null}],"processing_updated_time":{"$date":"2020-04-13T19:52:10.216Z"}}
+{"_id":"logkeeper","processing":true,"queue":[{"issue": "1234","patch_id": "526983de3ff1224641000002","source":"PR","enqueue_time": {"$date":"2020-04-13T19:52:00.182Z"},"modules": null}], "processing_updated_time":{"$date":"2020-04-13T19:52:10.216Z"}}
